+++ cmd/root.go
@@ -33,5 +33,6 @@ func init() {
 	rootCmd.AddCommand(initCmd)
 	rootCmd.AddCommand(planCmd)
 	rootCmd.AddCommand(updateCmd)
+	rootCmd.AddCommand(terraformCmd)
 	rootCmd.AddCommand(versionCmd)
 }