+++ config/parse_test.go
@@ -144,3 +144,67 @@ func TestParse(t *testing.T) {
 	}
 	assert.Equal(t, expected, config)
 }
+
+func TestParseMissingVars(t *testing.T) {
+
+	data := []byte(utils.TrimIndent(`
+        ---
+        mach_composer:
+          version: 1.0.0
+        global:
+          environment: test
+          terraform_config:
+            aws_remote_state:
+              bucket: "your bucket"
+              key_prefix: mach
+          cloud: aws
+        sites:
+        - identifier: my-site
+          endpoints:
+            main: api.my-site.nl
+            internal:
+              url: internal-api.my-site.nl
+              throttling_burst_limit: 5000
+              throttling_rate_limit: 10000
+          commercetools:
+            project_key: my-site
+            client_id: "<client-id>"
+            client_secret: "<client-secret>"
+            scopes: manage_api_clients:my-site manage_project:my-site view_api_clients:my-site
+            project_settings:
+              languages:
+                - en-GB
+                - nl-NL
+              currencies:
+                - GBP
+                - EUR
+              countries:
+                - GB
+                - NL
+          aws:
+            account_id: 123456789
+            region: eu-central-1
+          components:
+          - name: your-component
+            variables:
+              FOO_VAR: my-value
+			  BAR_VAR: ${var.foo}
+			  MULTIPLE_VARS: ${var.foo.bar} ${var.bar.foo}
+            secrets:
+              MY_SECRET: secretvalue
+        components:
+        - name: your-component
+          source: "git::https://github.com/<username>/<your-component>.git//terraform"
+          version: 0.1.0
+          endpoints:
+            internal: internal
+          integrations:
+            - aws
+            - commercetools
+    `))
+
+	// Empty variables, it should fail because var.foo cannot be resolved
+	vars := Variables{}
+	_, err := Parse(data, &vars)
+	assert.Error(t, err)
+}