+++ runner/apply.go
@@ -31,6 +31,20 @@ func TerraformApply(cfg *config.MachConfig, locations map[string]string, options
 	}
 }
 
+func TerraformProxy(cfg *config.MachConfig, locations map[string]string, siteName string, cmd []string) {
+	ctx := context.Background()
+
+	for i := range cfg.Sites {
+		site := cfg.Sites[i]
+
+		if siteName != "" && site.Identifier != siteName {
+			continue
+		}
+
+		RunTerraform(ctx, locations[site.Identifier], cmd...)
+	}
+}
+
 func TerraformApplySite(ctx context.Context, cfg *config.MachConfig, site *config.Site, path string, options *ApplyOptions) {
 
 	if !options.Reuse {
@@ -61,9 +75,9 @@ func TerraformApplySite(ctx context.Context, cfg *config.MachConfig, site *confi
 
 func TerraformPlanDetect(path string) string {
 	filename := GeneratePlanName(path)
-	filepath := filepath.Join(path, filename)
+	filePath := filepath.Join(path, filename)
 
-	if _, err := os.Stat(filepath); err == nil {
+	if _, err := os.Stat(filePath); err == nil {
 		return filename
 	}
 	return ""