+++ utils/exec_windows.go
@@ -4,6 +4,6 @@ package utils
 
 import "os/exec"
 
-func CmdSetForegrond(cmd *exec.Cmd) {
+func CmdSetForeground(cmd *exec.Cmd) {
 
 }