+++ CHANGELOG.md
@@ -5,6 +5,7 @@
   - Upgraded commercetools provider to 0.30.0
   - Upgraded Amplience provider to 0.3.7
   - Upgraded Azure provider to 2.99.0
+- Deprecate `commercetools.frontend` block, will be removed in a later release.
 
 ## 2.1.1 (2022-04-22)
 - Don't crash when running `mach-composer apply` without `--auto-approve`